+++ include/tao/pegtl/internal/missing_apply0.hpp
@@ -1,5 +1,6 @@
 // Copyright (c) 2019-2021 Dr. Colin Hirsch and Daniel Frey
-// Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
 #ifndef TAO_PEGTL_INTERNAL_MISSING_APPLY0_HPP
 #define TAO_PEGTL_INTERNAL_MISSING_APPLY0_HPP