+++ src/test/pegtl/verify_ifmt.hpp
@@ -1,5 +1,6 @@
 // Copyright (c) 2014-2021 Dr. Colin Hirsch and Daniel Frey
-// Please see LICENSE for license or visit https://github.com/taocpp/PEGTL/
+// Distributed under the Boost Software License, Version 1.0.
+// (See accompanying file LICENSE_1_0.txt or copy at https://www.boost.org/LICENSE_1_0.txt)
 
 #ifndef TAO_PEGTL_SRC_TEST_PEGTL_VERIFY_IFMT_HPP
 #define TAO_PEGTL_SRC_TEST_PEGTL_VERIFY_IFMT_HPP